requirements.txt
streamlit
torch
transformers
sentence-transformers
faiss-cpu
pypdf
Pillow
python-docx
